+++ activeuf/acquisition_function/__init__.py
@@ -1,0 +1,9 @@
+from activeuf.acquisition_function.dts import DoubleThompsonSampling
+from activeuf.acquisition_function.infomax import InfoMax
+from activeuf.acquisition_function.random import RandomAcquisitionFunction
+
+__all__ = [
+    "DoubleThompsonSampling",
+    "InfoMax",
+    "RandomAcquisitionFunction",
+]
