+++ activeuf/acquisition_function/random.py
@@ -1,7 +1,7 @@
 import numpy as np
 import torch
 
-from activeuf.acquisition_function.acquisition import BaseAcquisitionFunction
+from activeuf.acquisition_function.base import BaseAcquisitionFunction
 
 
 class RandomAcquisitionFunction(BaseAcquisitionFunction):
